new selectAssets func; new AssetSwap type;
